+++ plugins/accessToken/plugin.go
@@ -3,6 +3,36 @@
 
 package main
 
+import (
+	"context"
+	"fmt"
+	"github.com/go-redis/redis/v8"
+	"github.com/rs/zerolog/log"
+	plugins "isc-route-service/plugins/common"
+	"net/http"
+)
+
+var IscAccessTokenKey = "isc-access-token"
+var GatewayApiServiceAccessTokenRedisKeyPrefix = "gateway:api:service:access:token:"
+var ac = &AccessTokeConf{}
+
+type AccessTokeConf struct {
+	AccessToken struct {
+		urls []string `yaml:"urls"`
+	} `yaml:"access-token"`
+}
+
+var RedisClient redis.Client
+
+func init() {
+	//这里做初始化操作
+	plugins.ReadYamlToStruct("accessToken/conf.yml", ac)
+	if len(ac.AccessToken.urls) == 0 {
+		ac.AccessToken.urls = []string{"/api/common"}
+	}
+	RedisClient = *plugins.InitRedisClient("accessToken/conf.yml")
+}
+
 //Valid access token 验证
 func Valid(req *http.Request, target []byte) error {
 	uri := req.URL.Path
@@ -20,7 +50,7 @@ func Valid(req *http.Request, target []byte) error {
 		return err
 	} else {
 		//todo 需要连接redis
-		value := redisClient.Get(context.Background(), fmt.Sprintf("%s%s", GatewayApiServiceAccessTokenRedisKeyPrefix, iscAccessToken)).Val()
+		value := RedisClient.Get(context.Background(), fmt.Sprintf("%s%s", GatewayApiServiceAccessTokenRedisKeyPrefix, iscAccessToken)).Val()
 		if value == "" {
 			return err
 		}
